- Apparently, YASM is not a suitable substitute for NASM when doing Win32 builds. - Removed extraneous printf parameter for Texman.Init startup message. - Added newlines to the ends of a few headers that were missing them. - Fixed more GCC errors/warnings.
SVN r1232 (trunk)

+++ src/st_stuff.cpp
@@ -346,7 +346,7 @@ bool ST_Responder (event_t *ev)
 		static int counts[] = { countof(DoomCheats), countof(HereticCheats), countof(HexenCheats), 
 								countof(StrifeCheats), countof(ChexCheats), countof(SpecialCheats) };
 
-		for (int i=0; i<countof(cheatlists); i++)
+		for (size_t i=0; i<countof(cheatlists); i++)
 		{
 			if (CheatCheckList(ev, cheatlists[i], counts[i])) return true;
 		}
